requirements.txt
flet
pymongo
python-dotenv